sundials

Struct Cvode

pub struct Cvode {
    pub(crate) ptr: *mut c_void,
}

Fields
ptr: *mut c_void
Implementations

impl Cvode

pub fn new_bdf(ctx: &SunContext) -> Self

Create a new CVODE solver in BDF mode with the given context.

Panics

Panics if the solver creation fails (i.e., if the returned pointer is null which can happen if there is an issue with the context).

pub fn as_raw(&self) -> *mut c_void

pub fn set_userdata<T>(&self, data: &mut T)

pub fn init(&self, rhs: CVRhsFn, t0: f64, y: &NVector)

pub fn reinit(&self, t0: f64, y: &NVector)

pub fn set_tolerances(&self, reltol: f64, abstol: &NVector)

pub fn set_linear_solver(&self, linsol: &LinearSolver)

pub fn integrate( &self, tout: f64, y: &mut NVector, t: &mut f64, ) -> Result<(), CvodeError>

Integrate the ODE system up to time tout, storing the solution in y and the actual time reached in t.

Errors

Returns an error if the integration fails, which can happen for various reasons such as exceeding the maximum number of steps, convergence failures, or issues with the right-hand side function.

pub fn integrate_one_step(&self, tout: f64, y: &NVector, t: &mut f64) -> i32

pub fn get_current_time(&self) -> f64

pub fn get_last_step(&self) -> f64

pub fn get_current_step(&self) -> f64

pub fn set_preconditioner(&self, setup: CVLsPrecSetupFn, solve: CVLsPrecSolveFn)

pub fn set_wf_tolerances(&self, efun: CVEwtFn)

pub fn set_constraints(&self, constraints: &NVector)

pub fn set_max_nonlin_iters(&self, n: i32)

pub fn set_max_conv_fails(&self, n: i32)

pub fn set_eta_conv_fail(&self, eta: f64)

pub fn set_max_err_test_fails(&self, n: i32)

pub fn set_max_constraints_fails(&self, n: i32)

pub fn set_init_step(&self, h0: f64)

pub fn set_min_step(&self, hmin: f64)

pub fn set_max_step(&self, hmax: f64)

pub fn set_max_ord(&self, ord: i32)

pub fn set_max_num_steps(&self, n: i64)

pub fn set_max_hnil_warns(&self, n: i32)

pub fn set_stability_limit_detection(&self, enable: i32)

pub fn set_jac_eval_frequency(&self, freq: i64)

pub fn set_linear_solver_setup_frequency(&self, freq: i64)

pub fn set_epslin(&self, eps: f64)

pub fn save_statistics(&self, fname: &str)

Save CVODE statistics to a file named cvode_stats.txt in a human-readable table format.

Panics

Panics if the file cannot be opened for writing, which can happen due to permission issues or if the filesystem is read-only. Panics if the CVODE statistics cannot be printed, which can happen if there is an internal error in the CVODE library. Note that this function will overwrite the file cvode_stats.txt if it already exists.